Implement per-segment GroupAggr operator.
The semantics of GroupAggr changed to work per segment of the input
vector. This basically means that we have to include segment identifiers
in the grouping key. In the case of a flat vector consisting only of the
unit segment these segment identifiers are constant and can be easily
removed.