Fix rewrite for constant antijoin predicates.
Previously, the rewrite replaced antijoins with a constant predicate
that evaluates to true by a selection with a false predicate (i.e. an
empty relation). However, this is only correct if we can statically
determine that the right input is *not* empty. If it is empty, no tuple
from the left will find a match (regardless of the predicate) and thus
all left tuples survive the antijoin.

The same problem exists for the semijoin.